[REFACTOR] 스트림 이름 변경

+++ src/main/java/backend/globber/membertravel/controller/dto/CountryRecordDto.java
@@ -9,7 +9,7 @@ public record CountryRecordDto(
         String countryName,
         String countryCode,
         String continent,
-        int diaryCount,
+        long diaryCount,
         List<CityRecordDto> cities
 ) {
 }
